requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.0
pandas==2.1.3
python-dotenv==1.0.0
openai==1.3.5
jinja2==3.1.2
httpx==0.25.1
pydantic==2.5.1 